pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "whattsanddrops"
version = "0.1.0"
description = "A Python project"
requires-python = ">=3.8"
readme = "README.md"
dependencies = [
    "python-telegram-bot>=21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
]
